+++ lib/rundoc/code_command/website/screenshot.rb
@@ -11,10 +11,13 @@ class Rundoc::CodeCommand::Website
 
     def call(env = {})
       puts "Taking screenshot: #{@driver.current_url}"
-      filename = @driver.screenshot(upload: @upload, screenshots_path: env[:screenshots_path])
+      filename = @driver.screenshot(
+        upload: @upload,
+        screenshots_dir: env[:context].screenshots_dir
+      )
 
-      relative_filename = filename.relative_path_from(env[:output_dir])
-      env[:replace] = "![Screenshot of #{@driver.current_url}](#{relative_filename})"
+      relative_filename = filename.relative_path_from(env[:context].output_dir)
+      env[:before] << "![Screenshot of #{@driver.current_url}](#{relative_filename})"
       ""
     end